A bit of code cleanup and formatting.

+++ adventurebackpack/blocks/BlockSleepingBag.java
@@ -33,7 +33,7 @@ public class BlockSleepingBag extends BlockContainer {
 		setUnlocalizedName(BlockInfo.SLEEPINGBAG_NAME);
 	}
 
-	private int[][] footBlockToHeadBlockMap = new int[][] { { 0, 1 },{ -1, 0 }, { 0, -1 }, { 1, 0 } };
+	private int[][] footBlockToHeadBlockMap = new int[][] { { 0, 1 }, { -1, 0 }, { 0, -1 }, { 1, 0 } };
 
 	@Override
 	public int getRenderType() {
@@ -56,15 +56,16 @@ public class BlockSleepingBag extends BlockContainer {
 	}
 
 	@Override
-	public void onBlockPlacedBy(World world, int x, int y, int z,EntityLivingBase player, ItemStack stack) {
-		int dir = MathHelper.floor_double((double) ((player.rotationYaw * 4F) / 360F) + 0.5D) & 3;
+	public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase player, ItemStack stack) {
+		int dir = MathHelper.floor_double((player.rotationYaw * 4F) / 360F + 0.5D) & 3;
 		world.setBlockMetadataWithNotify(x, y, z, dir, 3);
 		createNewTileEntity(world);
 	}
 
 	@Override
-	public boolean canPlaceBlockOnSide(World par1World, int par2, int par3,int par4, int side) {
-		if (ForgeDirection.getOrientation(side) == ForgeDirection.UP)return true;
+	public boolean canPlaceBlockOnSide(World par1World, int par2, int par3, int par4, int side) {
+		if (ForgeDirection.getOrientation(side) == ForgeDirection.UP)
+			return true;
 		return false;
 	}
 
@@ -94,12 +95,13 @@ public class BlockSleepingBag extends BlockContainer {
 	}
 
 	@Override
-	public boolean onBlockActivated(World world, int x, int y, int z,EntityPlayer player, int par6, float par7, float par8, float par9) {
+	public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer player, int par6, float par7,
+			float par8, float par9) {
 		if (world.isRemote) {
 			return true;
 		} else {
 			int meta = world.getBlockMetadata(x, y, z);
-			
+
 			if (world.provider.canRespawnHere() && world.getBiomeGenForCoords(x, z) != BiomeGenBase.hell) {
 				if (isBedOccupied(meta)) {
 					EntityPlayer entityplayer1 = null;
@@ -111,9 +113,7 @@ public class BlockSleepingBag extends BlockContainer {
 						if (entityplayer2.isPlayerSleeping()) {
 							ChunkCoordinates chunkcoordinates = entityplayer2.playerLocation;
 
-							if (chunkcoordinates.posX == x
-									&& chunkcoordinates.posY == y
-									&& chunkcoordinates.posZ == z) {
+							if (chunkcoordinates.posX == x && chunkcoordinates.posY == y && chunkcoordinates.posZ == z) {
 								entityplayer1 = entityplayer2;
 							}
 						}
@@ -142,9 +142,9 @@ public class BlockSleepingBag extends BlockContainer {
 					return true;
 				}
 			} else {
-				double d0 = (double) x + 0.5D;
-				double d1 = (double) y + 0.5D;
-				double d2 = (double) z + 0.5D;
+				double d0 = x + 0.5D;
+				double d1 = y + 0.5D;
+				double d2 = z + 0.5D;
 				world.setBlockToAir(x, y, z);
 				int k1 = getDirection(meta);
 				x += footBlockToHeadBlockMap[k1][0];
@@ -152,14 +152,12 @@ public class BlockSleepingBag extends BlockContainer {
 
 				if (world.getBlockId(x, y, z) == this.blockID) {
 					world.setBlockToAir(x, y, z);
-					d0 = (d0 + (double) x + 0.5D) / 2.0D;
-					d1 = (d1 + (double) y + 0.5D) / 2.0D;
-					d2 = (d2 + (double) z + 0.5D) / 2.0D;
+					d0 = (d0 + x + 0.5D) / 2.0D;
+					d1 = (d1 + y + 0.5D) / 2.0D;
+					d2 = (d2 + z + 0.5D) / 2.0D;
 				}
 
-				world.newExplosion((Entity) null, (double) ((float) x + 0.5F),
-						(double) ((float) y + 0.5F),
-						(double) ((float) z + 0.5F), 5.0F, true, true);
+				world.newExplosion((Entity) null, x + 0.5F, y + 0.5F, z + 0.5F, 5.0F, true, true);
 				return true;
 			}
 		}
@@ -179,14 +177,14 @@ public class BlockSleepingBag extends BlockContainer {
 	}
 
 	@Override
-	public void setBedOccupied(World world, int x, int y, int z,EntityPlayer player, boolean occupied) {
+	public void setBedOccupied(World world, int x, int y, int z, EntityPlayer player, boolean occupied) {
 		int meta = world.getBlockMetadata(x, y, z);
 		meta = occupied ? meta |= 4 : meta & -5;
 		world.setBlockMetadataWithNotify(x, y, z, meta, 4);
 	}
 
 	@Override
-	public boolean isBed(World world, int x, int y, int z,EntityLivingBase player) {
+	public boolean isBed(World world, int x, int y, int z, EntityLivingBase player) {
 		return true;
 	}
 
@@ -196,14 +194,14 @@ public class BlockSleepingBag extends BlockContainer {
 	}
 
 	@Override
-	public boolean isBlockSolidOnSide(World world, int x, int y, int z,ForgeDirection side) {
+	public boolean isBlockSolidOnSide(World world, int x, int y, int z, ForgeDirection side) {
 		return false;
 	}
 
 	@Override
-	public void setBlockBoundsBasedOnState(IBlockAccess blockAccess, int x,int y, int z) {
+	public void setBlockBoundsBasedOnState(IBlockAccess blockAccess, int x, int y, int z) {
 		int orientation = blockAccess.getBlockMetadata(x, y, z);
-		switch(orientation){
+		switch (orientation) {
 		case 0:
 			setBlockBounds(0.0F, 0.0F, 0.0F, 1.0F, 0.2F, 2.0F);
 			break;
@@ -219,20 +217,19 @@ public class BlockSleepingBag extends BlockContainer {
 		default:
 			break;
 		}
-		
-	
+
 	}
-	
+
 	@Override
 	public AxisAlignedBB getCollisionBoundingBoxFromPool(World world, int x, int y, int z) {
 		setBlockBoundsBasedOnState(world, x, y, z);
-		return AxisAlignedBB.getAABBPool().getAABB((double)x + this.minX, (double)y + this.minY, (double)z + this.minZ, (double)x + this.maxX, (double)y + this.maxY, (double)z + this.maxZ);
-	}	
-	
+		return AxisAlignedBB.getAABBPool().getAABB(x + this.minX, y + this.minY, z + this.minZ, x + this.maxX,
+				y + this.maxY, z + this.maxZ);
+	}
 
 	@Override
 	public MovingObjectPosition collisionRayTrace(World world, int x, int y, int z, Vec3 start, Vec3 end) {
 		setBlockBoundsBasedOnState(world, x, y, z);
 		return super.collisionRayTrace(world, x, y, z, start, end);
-	}		
+	}
 }
